fixed DRP, fixed stringtables
