+++ addons/backpacks/stringtable.xml
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project name="ttt_a3">
-    <Package name="backpacks"/>
+    <Package name="backpacks">
         <Key ID="STR_ttt_backpacks_radio_coyote">
             <English>TTT Radio Backpack (Coyote)</English>
             <German>TTT Funk Rucksack (Coyote)</German>
         </Key>
+    </Package>
 </Project>
