+++ addons/compat_3cb_vehicles/stringtable.xml
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project name="ttt_a3">
-    <Package name="compat_3cb_vehicles"/>
+    <Package name="compat_3cb_vehicles">
         <Key ID="STR_ttt_compat_cup_vehicles_takegear">
             <English>Take gear</English>
             <German>Ausrüstung Nehmen</German>
         </Key>
+    </Package>
 </Project>
