+++ addons/compat_rnt/stringtable.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project name="ttt_a3">
-    <Package name="compat_rnt"/>
+    <Package name="compat_rnt">
+    
 </Project>
