+++ addons/discord_richpresence/stringtable.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project name="ttt_mod">
-    <Package name="discord_richpresence"/>
+    <Package name="discord_richpresence">
+    </Package>
 </Project>
