+++ addons/helmets/stringtable.xml
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project name="ttt_a3">
-    <Package name="helmets"/>
+    <Package name="helmets">
         <Key ID="STR_ttt_helmets_buzzard_displayName">
             <English>TTT Helicopter Helmet Buzzard</English>
             <German>TTT Helikopter Helm Bussard</German>
         </Key>
+    </Package>
 </Project>
