+++ addons/main/stringtable.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project name="ttt_a3">
-    <Package name="main"/>
+    <Package name="main">
         <Key ID="STR_ttt_main_URL">
             <English>https://tacticalteam.de</English>
             <German>https://tacticalteam.de</German>
@@ -9,4 +9,5 @@
             <English>Tactical Training Team</English>
             <German>Tactical Training Team</German>
         </Key>
+    </Package>
 </Project>
