+++ addons/rsr_core/stringtable.xml
@@ -9,4 +9,5 @@
             <English>Redd'n'Tank</English>
             <German>Redd'n'Tank</German>
         </Key>
+    </Package>
 </Project>
