+++ addons/rsr_itc_compat/stringtable.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
  <Project name="ttt_a3">
-    <Package name="rsr_itc_compat"/>
+    <Package name="rsr_itc_compat">
+    </Package>
 </Project>
