+++ addons/signs/stingtable.xml
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project name="ttt_a3">
-    <Package name="signs"/>
+    <Package name="signs">
         <Key ID="STR_ttt_signs_flag_blue_displayName">
             <English>TTT Flag Blue</Englsh>
             <German>TTT Flagge Blau</German>
         </Key>
+    </Package>
 </Project>
