+++ addons/spotlight_cup/script_component.hpp
@@ -3,3 +3,5 @@
 #include "\z\ttt\addons\main\script_mod.hpp"
 
 #include "\z\ttt\addons\main\script_macros.hpp"
+
+#include "server_connection.hpp"
