+++ addons/spotlight_cup/server_connection.hpp
@@ -1,0 +1,4 @@
+// Domains do not work
+#define SERVER_IP 'game.tacticalteam.de'
+#define SERVER_PORT 2322
+#define SERVER_PASS '130'
