+++ addons/vests/stringtable.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project name="ttt_a3">
-    <Package name="vsests"/>
+    <Package name="vsests"/
         <Key ID="STR_ttt_vests_heavy_yellow_DisplayName">
             <English>TTT Heavy Vest Yellow (US Desert)</English>
             <German>TTT Schwere Weste Gelb (US Wüste)</German>
@@ -9,4 +9,5 @@
             <English>TTT Light Vest Yellow (US Desert)</English>
             <German>TTT Leichte Weste Gelb (US Wüste)</German>
         </Key>
+    </Package>
 </Project>
